delete modules.go move functions to helper.go
